+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### :house: Internal
 - `examples`
   - [#850](https://github.com/wix-incubator/rich-content/pull/850) fix performance
+  - [#850](https://github.com/wix-incubator/rich-content/pull/850) save content to local storage
 
 </details>
 <hr/>